refactor(store): Rename stores to store
