+++ src/components/core/LayoutView/script.js
@@ -3,7 +3,7 @@ import { mapActions, mapState } from 'vuex';
 import VtkView from 'paraview-glance/src/components/core/VtkView';
 import viewHelper from 'paraview-glance/src/components/core/VtkView/helper';
 import { Breakpoints } from 'paraview-glance/src/constants';
-import { Actions, Getters, Mutations } from 'paraview-glance/src/stores/types';
+import { Actions, Getters, Mutations } from 'paraview-glance/src/store/types';
 
 // ----------------------------------------------------------------------------
 // Component API
