+++ src/store/globalSettings.js
