+++ src/store/views.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 
 import { VIEW_TYPES } from 'paraview-glance/src/components/core/VtkView/constants';
 import viewHelper from 'paraview-glance/src/components/core/VtkView/helper';
-import { Actions, Getters, Mutations } from 'paraview-glance/src/stores/types';
+import { Actions, Getters, Mutations } from 'paraview-glance/src/store/types';
 
 function setViewData(state, viewId, data) {
   Vue.set(
